backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
ebooklib==0.18
PyPDF2==3.0.1
pdfplumber==0.10.3
beautifulsoup4==4.12.2
sqlalchemy==2.0.23
psycopg2-binary>=2.9.9
openai>=1.12.0
google-generativeai>=0.3.0
httpx>=0.27.0
python-dotenv==1.0.0
requests>=2.31.0
